+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Added asUnmodifiableMap() method to the ReadableMapProperty interface
 * Added interfaces and abstract implementations of a set property
 * Fixed type erasure problem of the remove() method in the SimpleListProperty
+* Fixed a possible IllegalStateException in the AbstractComponentDecoration when calculating the clipping bounds
 * Added JButtonMnemonicProperty and JLabelDisplayedMnemonicProperty
 * Added ActionProperty
 * Added convenient getter and clear methods in the CompositeReadableProperty and CompositeWritable properties